requirements.txt
# ============================================
# AI Agent 健康数据分析项目依赖
# ============================================

langchain-community>=0.0.10  # 关键：包含ChatOllama

# ========== 核心AI Agent依赖 ==========
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.6.0

# ========== 数据处理 ==========
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# ========== 可视化 ==========
matplotlib>=3.7.0
seaborn>=0.13.0

# ========== 配置管理 ==========
pyyaml>=6.0
python-dotenv>=1.0.0

# ========== 开发工具 ==========
jupyter>=1.0.0
ipython>=8.14.0
black>=23.0.0
flake8>=6.0.0

# ========== 测试 ==========
pytest>=7.4.0
pytest-cov>=4.1.0

# ========== 可选（高级功能） ==========
streamlit>=1.28.0  # 如果要做Web界面
fastapi>=0.104.0  # 如果要做API
pydantic>=2.0.0   # 数据验证